Adapt hash router spec for immutable
